chatbot/requirements.txt
chromadb
sentence-transformers
boto3
PyPDF2
python-docx
